machine_monitor/backend/requirements.txt
setuptools>=65.5.1
wheel
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
pydantic-settings==2.0.1
sqlalchemy==2.0.10
websockets==12.0
python-multipart==0.0.9
prometheus-client==0.19.0
kafka-python==2.0.2
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.26.0
